third_party/blink/web_tests/external/wpt/css/css-color/parsing/opacity-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['opacity'] = "clamp(50%, 0%, 70%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0.5)" but got "clamp(50%, 0%, 70%)"
[FAIL] e.style['opacity'] = "clamp(50%, 80%, 70%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0.7)" but got "clamp(50%, 80%, 70%)"
[FAIL] e.style['opacity'] = "clamp(50%, 60%, 70%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0.6)" but got "clamp(50%, 60%, 70%)"
[FAIL] e.style['opacity'] = "min(50%, 0%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0)" but got "min(50%, 0%)"
[FAIL] e.style['opacity'] = "min(0%, 50%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0)" but got "min(0%, 50%)"
[FAIL] e.style['opacity'] = "max(50%, 0%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0.5)" but got "max(50%, 0%)"
[FAIL] e.style['opacity'] = "max(0%, 50%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(0.5)" but got "max(0%, 50%)"
[FAIL] e.style['opacity'] = "min(-40%, 50%)" should set the property value
  assert_equals: serialization should be canonical expected "calc(-0.4)" but got "min(-40%, 50%)"
Harness: the test ran to completion.
